Update the product removal command

+++ Product.Application/Features/Commands/Handlers/DeleteProductDtoRequestHandler.cs
@@ -1,24 +1,54 @@
-﻿using AutoMapper;
-using MediatR;
+﻿using MediatR;
+using Product.Application.DTOs.Validator;
+using Product.Application.Exceptions;
 using Product.Application.Features.Requests.Handlers;
 using Product.Application.Interfaces;
+using Product.Application.Response;
 
 namespace Product.Application.Features.Commands.Handlers
 {
-    public class DeleteProductDtoRequestHandler(IProductDbContext context, IMapper mapper) : IRequestHandler<DeleteProductDtoRequest, Unit>
+    public class DeleteProductDtoRequestHandler(IProductDbContext context, IDeleteProductDtoValidator deleteValidator) : IRequestHandler<DeleteProductDtoRequest, ProductAPIResponse>
     {
         private readonly IProductDbContext _context = context;
-        private readonly IMapper _mapper = mapper;
+        private readonly IDeleteProductDtoValidator _deleteValidator = deleteValidator;
+        private readonly ProductAPIResponse _productAPIResponse = new ProductAPIResponse();
 
-        public async Task<Unit> Handle(DeleteProductDtoRequest request, CancellationToken cancellationToken)
+        public async Task<ProductAPIResponse> Handle(DeleteProductDtoRequest request, CancellationToken cancellationToken)
         {
-            var product = await _context.Products.FindAsync(new object[] { request.DeleteProduct.Id }, cancellationToken)
-                ?? throw new Exception($"Продукт с идентификатором ({request.DeleteProduct.Id}) не существует");
+            try
+            {
+                var validator = await _deleteValidator.ValidateAsync(request.DeleteProduct, cancellationToken);
 
-            _context.Products.Remove(product);
-            await _context.SaveChangesAsync(cancellationToken);
+                if (!validator.IsValid)
+                {
+                    _productAPIResponse.IsSuccess = false;
+                    _productAPIResponse.Message = "Ошибка удаления продукта";
+                    _productAPIResponse.ValidationErrors = validator.Errors.Select(x => x.ErrorMessage).ToList();
+                }
 
-            return Unit.Value;
+                else
+                {
+                    var product = await _context.Products.FindAsync(new object[] { request.DeleteProduct.Id }, cancellationToken)
+                        ?? throw new NotFoundException($"Продукт c идентификатором ({request.DeleteProduct.Id}) не существует", request.DeleteProduct.Id);
+
+                    _context.Products.Remove(product);
+                    await _context.SaveChangesAsync(cancellationToken);
+
+                    _productAPIResponse.IsSuccess = true;
+                    _productAPIResponse.Message = "Продукт успешно удален";
+                    _productAPIResponse.Id = product.Id;
+
+                    return _productAPIResponse;
+                }
+            }
+
+            catch (Exception exception)
+            {
+                _productAPIResponse.IsSuccess = false;
+                _productAPIResponse.Message = exception.Message;
+            }
+
+            return _productAPIResponse;
         }
     }
 }
